multilingual-translator/requirements.txt
streamlit>=1.32.0
openai-whisper>=20231117
transformers>=4.38.0
torch>=2.0.0
sounddevice>=0.4.6
scipy>=1.12.0
gtts>=2.5.0
numpy>=1.24.0
noisereduce>=3.0.0
sentencepiece>=0.1.99
protobuf>=3.20.0
